Revert a change that can't work in UE4.

+++ Source/CesiumRuntime/Public/CesiumCamera.h
@@ -41,7 +41,7 @@ public:
    * @brief The horizontal field of view of the camera in degrees.
    */
   UPROPERTY(BlueprintReadWrite, Category = "Cesium")
-  double FieldOfViewDegrees;
+  float FieldOfViewDegrees;
 
   /**
    * @brief The overriden aspect ratio for this camera.
@@ -53,7 +53,7 @@ public:
    * within a larger viewport.
    */
   UPROPERTY(BlueprintReadWrite, Category = "Cesium")
-  double OverrideAspectRatio = 0.0f;
+  float OverrideAspectRatio = 0.0f;
 
   /**
    * @brief Construct an uninitialized FCesiumCamera object.
@@ -72,7 +72,7 @@ public:
       const FVector2D& ViewportSize,
       const FVector& Location,
       const FRotator& Rotation,
-      double FieldOfViewDegrees);
+      float FieldOfViewDegrees);
 
   /**
    * @brief Construct a new FCesiumCamera object.
@@ -87,6 +87,6 @@ public:
       const FVector2D& ViewportSize,
       const FVector& Location,
       const FRotator& Rotation,
-      double FieldOfViewDegrees,
-      double OverrideAspectRatio);
+      float FieldOfViewDegrees,
+      float OverrideAspectRatio);
 };
